requirements.txt
anyio==4.4.0
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.3.1
faster-whisper==1.0.3
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
ollama==0.3.2
onnxruntime==1.19.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.28.0
PyAudio==0.2.14
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
sympy==1.13.2
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
